requirements.txt
# requirements.txt
fastapi
uvicorn
requests
websockets
python-dotenv
google-cloud-speech
